# OpenAPI/pokedex.yaml
openapi: 3.0.0
info:
  title: Pokémon API
  version: 1.0.0
  description: API to retrieve, sort, filter, and add Pokémon data.
paths:
  /data:
    get:
      summary: Retrieve all Pokémon data
      parameters:
        - name: sort_by
          in: query
          description: Comma-separated list of columns to sort by.
          required: false
          schema:
            type: string
        - name: filter
          in: query
          description: Filter criteria in the format field=value.
          required: false
          explode: true
          style: form
          schema:
            type: string
      responses:
        '200':
          description: A list of Pokémon
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    Image:
                      type: string
                    Index:
                      type: integer
                    Name:
                      type: string
                    Type 1:
                      type: string
                    Type 2:
                      type: string
                    Total:
                      type: integer
                    HP:
                      type: integer
                    Attack:
                      type: integer
                    Defense:
                      type: integer
                    SP. Atk.:
                      type: integer
                    SP. Def:
                      type: integer
                    Speed:
                      type: integer
    post:
      summary: Add a single Pokémon record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Image:
                  type: string
                Index:
                  type: integer
                Name:
                  type: string
                Type 1:
                  type: string
                Type 2:
                  type: string
                Total:
                  type: integer
                HP:
                  type: integer
                Attack:
                  type: integer
                Defense:
                  type: integer
                SP. Atk.:
                  type: integer
                SP. Def:
                  type: integer
                Speed:
                  type: integer
      responses:
        '201':
          description: Pokémon record added
  /data/batch:
    post:
      summary: Add multiple Pokémon records
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  Image:
                    type: string
                  Index:
                    type: integer
                  Name:
                    type: string
                  Type 1:
                    type: string
                  Type 2:
                    type: string
                  Total:
                    type: integer
                  HP:
                    type: integer
                  Attack:
                    type: integer
                  Defense:
                    type: integer
                  SP. Atk.:
                    type: integer
                  SP. Def:
                    type: integer
                  Speed:
                    type: integer
      responses:
        '201':
          description: Pokémon records added